move mapColors.ts to components/colors.ts
