+++ src/components/BottomDrawer/index.tsx
@@ -19,11 +19,14 @@ import {
 import { Dataset } from "@/types/api";
 import { useLayerStore } from "@/store/layer-store";
 import { getAttributes } from "@/utils/getAttributes";
-import { consolidateTimeSeries, hasMonthlyVariation } from "@/utils/consolidateTimeSeries";
+import {
+  consolidateTimeSeries,
+  hasMonthlyVariation,
+} from "@/utils/consolidateTimeSeries";
 import { formatYAxisLabel } from "@/utils/formatCharts";
 import { useAreaStore } from "@/store/area-store";
 import { useUiStore } from "@/store/ui-store";
-import { chartColors } from "../Map/mapColors";
+import { chartColors } from "../colors";
 
 const BottomDrawer = () => {
   const { layers, tabularLayerData, getLayerMetadata } = useLayerStore();
@@ -53,7 +56,10 @@ const BottomDrawer = () => {
   const showMonthlyView = hasMonthlyData && viewMode === "monthly";
 
   // Get time series data - tabularLayerData is already filtered by selected area
-  const timeSeriesData = consolidateTimeSeries(tabularLayerData, showMonthlyView);
+  const timeSeriesData = consolidateTimeSeries(
+    tabularLayerData,
+    showMonthlyView,
+  );
   const series = getAttributes(tabularLayerData).map((i, index) => ({
     name: i,
     color: `${index < chartColors.length ? chartColors[index] : "yellow"}.solid`,
@@ -73,7 +79,8 @@ const BottomDrawer = () => {
       collapsible
       variant="plain"
       value={isTimeSeriesOpen ? ["bottom-drawer"] : []}
-      onValueChange={(details) => setTimeSeriesOpen(details.value.includes("bottom-drawer"))}
+      onValueChange={(details) =>
+        setTimeSeriesOpen(details.value.includes("bottom-drawer"))}
       disabled={!tabularLayerData.length}
     >
       <Accordion.Item
@@ -107,9 +114,15 @@ const BottomDrawer = () => {
         <Accordion.ItemContent>
           <Accordion.ItemBody>
             <Box p={3}>
-              {isLoading ? 
-                <Skeleton height={4} width="40%" mb={4} loading={isLoading} /> :
-                <Box display="flex" justifyContent="space-between" alignItems="center" mb={4}>
+              {isLoading ? (
+                <Skeleton height={4} width="40%" mb={4} loading={isLoading} />
+              ) : (
+                <Box
+                  display="flex"
+                  justifyContent="space-between"
+                  alignItems="center"
+                  mb={4}
+                >
                   <Heading as="h4" fontSize="md">
                     {ac ? ac : province ? province : "National Level"} -{" "}
                     {layerMetadata
@@ -132,7 +145,8 @@ const BottomDrawer = () => {
                       </Button>
                     </ButtonGroup>
                   )}
-                </Box>}
+                </Box>
+              )}
               <Skeleton height="100%" loading={isLoading}>
                 <Chart.Root maxH="200px" chart={chart}>
                   <LineChart data={chart.data}>
@@ -151,8 +165,7 @@ const BottomDrawer = () => {
                       tickLine={false}
                       tickMargin={10}
                       type="number"
-                      tickFormatter={(value: number) =>
-                        String(formatYAxisLabel(value))}
+                      tickFormatter={(value: number) => formatYAxisLabel(value)}
                       stroke={chart.color("border")}
                     />
                     <Tooltip
